Fix Windows tray ticket submission
